+++ yandex/data_source_yandex_alb_load_balancer.go
@@ -351,6 +351,10 @@ func dataSourceHTTPHandler() *schema.Schema {
 					Type:     schema.TypeBool,
 					Optional: true,
 				},
+				"rewrite_request_id": {
+					Type:     schema.TypeBool,
+					Computed: true,
+				},
 			},
 		},
 	}