+++ yandex/data_source_yandex_alb_load_balancer_test.go
@@ -251,6 +251,69 @@ func TestAccDataSourceALBLoadBalancer_httpListenerWithAllowHTTP10(t *testing.T) 
 	})
 }
 
+func TestAccDataSourceALBLoadBalancer_httpListenerWithRewriteRequestID(t *testing.T) {
+	t.Parallel()
+
+	albResource := albLoadBalancerInfo()
+	albResource.IsHTTPListener = true
+	albResource.IsHTTPHandler = true
+	albResource.IsRewriteRequestID = true
+	albResource.IsDataSource = true
+
+	var alb apploadbalancer.LoadBalancer
+	listenerPath := ""
+
+	resource.Test(t, resource.TestCase{
+		PreCheck:     func() { testAccPreCheck(t) },
+		Providers:    testAccProviders,
+		CheckDestroy: testAccCheckALBLoadBalancerDestroy,
+		Steps: []resource.TestStep{
+			{
+				Config: testALBLoadBalancerConfig_basic(albResource),
+				Check: resource.ComposeTestCheckFunc(
+					testAccCheckALBLoadBalancerExists(albLoadBalancerDataSourceResource, &alb),
+					testExistsFirstElementWithAttr(
+						albLoadBalancerDataSourceResource, "listener", "name", &listenerPath,
+					),
+					testCheckResourceSubAttrFn(
+						albLoadBalancerDataSourceResource, &listenerPath, "name", func(value string) error {
+							albName := alb.GetListeners()[0].GetName()
+							if value != albName {
+								return fmt.Errorf("ALB Load Balancer's listener's name doesnt't match. %s != %s", value, albName)
+							}
+							return nil
+						},
+					),
+					testExistsFirstElementWithAttr(
+						albLoadBalancerDataSourceResource, "listener", "endpoint.0.ports.0", &listenerPath,
+					),
+					testCheckResourceSubAttrFn(
+						albLoadBalancerDataSourceResource, &listenerPath, "endpoint.0.ports.0", func(value string) error {
+							port := alb.GetListeners()[0].GetEndpoints()[0].GetPorts()[0]
+							if realValue, _ := strconv.ParseInt(value, 10, 64); realValue != port {
+								return fmt.Errorf("ALB Load Balancer's listener's endpoint's port doesnt't match. %d != %d", realValue, port)
+							}
+							return nil
+						},
+					),
+					testExistsFirstElementWithAttr(
+						albLoadBalancerDataSourceResource, "listener", "http.0.handler.0.rewrite_request_id", &listenerPath,
+					),
+					testCheckResourceSubAttrFn(
+						albLoadBalancerDataSourceResource, &listenerPath, "http.0.handler.0.rewrite_request_id", func(value string) error {
+							rewriteRequestID := alb.GetListeners()[0].GetHttp().GetHandler().GetRewriteRequestId()
+							if realValue, _ := strconv.ParseBool(value); realValue != rewriteRequestID {
+								return fmt.Errorf("ALB Load Balancer's HTTP listener's rewrite request id doesnt't match. %t != %t", realValue, rewriteRequestID)
+							}
+							return nil
+						},
+					),
+				),
+			},
+		},
+	})
+}
+
 func TestAccDataSourceALBLoadBalancer_httpListenerWithRedirects(t *testing.T) {
 	t.Parallel()
 
@@ -473,6 +536,69 @@ func TestAccDataSourceALBLoadBalancer_tlsListenerWithAllowHTTP10(t *testing.T) {
 							allowHTTP1 := alb.GetListeners()[0].GetTls().GetDefaultHandler().GetHttpHandler().GetAllowHttp10()
 							if realValue, _ := strconv.ParseBool(value); realValue != allowHTTP1 {
 								return fmt.Errorf("ALB Load Balancer's TLS listener's allow HTTP 1.0 doesnt't match. %t != %t", realValue, allowHTTP1)
+							}
+							return nil
+						},
+					),
+				),
+			},
+		},
+	})
+}
+
+func TestAccDataSourceALBLoadBalancer_tlsListenerRewriteRequestID(t *testing.T) {
+	t.Parallel()
+
+	albResource := albLoadBalancerInfo()
+	albResource.IsTLSListener = true
+	albResource.IsHTTPHandler = true
+	albResource.IsRewriteRequestID = true
+	albResource.IsDataSource = true
+
+	var alb apploadbalancer.LoadBalancer
+	listenerPath := ""
+
+	resource.Test(t, resource.TestCase{
+		PreCheck:     func() { testAccPreCheck(t) },
+		Providers:    testAccProviders,
+		CheckDestroy: testAccCheckALBLoadBalancerDestroy,
+		Steps: []resource.TestStep{
+			{
+				Config: testALBLoadBalancerConfig_basic(albResource),
+				Check: resource.ComposeTestCheckFunc(
+					testAccCheckALBLoadBalancerExists(albLoadBalancerDataSourceResource, &alb),
+					testExistsFirstElementWithAttr(
+						albLoadBalancerDataSourceResource, "listener", "name", &listenerPath,
+					),
+					testCheckResourceSubAttrFn(
+						albLoadBalancerDataSourceResource, &listenerPath, "name", func(value string) error {
+							albName := alb.GetListeners()[0].GetName()
+							if value != albName {
+								return fmt.Errorf("ALB Load Balancer's listener's name doesnt't match. %s != %s", value, albName)
+							}
+							return nil
+						},
+					),
+					testExistsFirstElementWithAttr(
+						albLoadBalancerDataSourceResource, "listener", "endpoint.0.ports.0", &listenerPath,
+					),
+					testCheckResourceSubAttrFn(
+						albLoadBalancerDataSourceResource, &listenerPath, "endpoint.0.ports.0", func(value string) error {
+							port := alb.GetListeners()[0].GetEndpoints()[0].GetPorts()[0]
+							if realValue, _ := strconv.ParseInt(value, 10, 64); realValue != port {
+								return fmt.Errorf("ALB Load Balancer's listener's endpoint's port doesnt't match. %d != %d", realValue, port)
+							}
+							return nil
+						},
+					),
+					testExistsFirstElementWithAttr(
+						albLoadBalancerDataSourceResource, "listener", "tls.0.default_handler.0.http_handler.0.rewrite_request_id", &listenerPath,
+					),
+					testCheckResourceSubAttrFn(
+						albLoadBalancerDataSourceResource, &listenerPath, "tls.0.default_handler.0.http_handler.0.rewrite_request_id", func(value string) error {
+							rewriteRequestID := alb.GetListeners()[0].GetTls().GetDefaultHandler().GetHttpHandler().GetRewriteRequestId()
+							if realValue, _ := strconv.ParseBool(value); realValue != rewriteRequestID {
+								return fmt.Errorf("ALB Load Balancer's TLS listener's rewrite request id doesnt't match. %t != %t", realValue, rewriteRequestID)
 							}
 							return nil
 						},